backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
httpx
requests
sqlalchemy
pandas
scikit-learn
shap
opencv-python
pillow
numpy
python-multipart
joblib
xgboost
torch
torchvision
torchaudio
